Route /objects gaat naar naar entities
